Add arm-trusted-firmware repository
This patch adds the arm-trusted-firmware repository. This is not a
repository for local development, but instead supposed to be a direct
mirror of the upstream github.com/ARM-software/arm-trusted-firmware.
It's cros/master branch will be automatically synced to the upstream
master through chrome-infra magic and no patches for it should be
uploaded through Gerrit. Firmware branches will still work as normal.

BUG=chromium:489876
TEST=Made equivalent change to ~/trunk/.repo/manifests, repo synced,
repository showed up and was checked out to the right commit.

Change-Id: I273da4a9f7d7a3b242aebc225f99beb2ae3f2e94
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273584
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>