Add arm-trusted-firmware repo to branch.
CL:273584 added the arm-trusted-firmware to the manifest. This branch
was cut before that CL, but we'd like to include the project on this
branch.

This is not a simple cherry-pick from master, because branch-util
modifies the manifest after branching, leading to conflicts.

BUG=chrome-os-partner:41742
TEST='repo sync -m test' and check that arm-trusted-firmware is sync'ed.

Change-Id: I9f909d1f275a29de95d711143138faba130ca4aa
Reviewed-on: https://chromium-review.googlesource.com/282625
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>