manifest: Remove oauth2client and upgrade swarming
It turns out swarming.client already depends on oauth2client, so the
extra oauth2client in chromite/third_party is superfluous and can
conflict with it if the versions differ.

The ts-mon library requires oauth2client to be at a newer version (with
the GoogleCredentials class), so this change upgrades swarming.client to
a revision which has the newer oauth2client.

BUG=chromium:583705
TEST=None
CQ-DEPEND=CL:*247148

Change-Id: Ibe670334cb6b7dfc26eb815c01579fa03a0f8cfa
Reviewed-on: https://chromium-review.googlesource.com/325820
Commit-Ready: Paul Hobbs <phobbs@google.com>
Tested-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Paul Hobbs <phobbs@google.com>