arm-trusted-firmware: Uprev pinned HEAD and set explicit branch-mode
This patch uprevs the pinned HEAD for the arm-trusted-firmware
repository to the newest upstream master commit "Merge pull request #546
from mtk09422/mtk-bl31-update".

Also explicitly add the branch-mode="create" annotation. Even though
this mode is already the default for chromiumos/* repositories, adding
this communicates more clearly that this is a desired and important
property for this repository.

CQ-DEPEND=CL:*251132
BUG=None
TEST=None (that's what we have a CQ for, right?)

Change-Id: I8215d0963ed2f4b89cd178264810f9e366cd3542
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332249
Reviewed-by: Don Garrett <dgarrett@chromium.org>