manifest: Pin GCC aosp to a particular commit.
We don't want the GCC compiler in ChromeOS changing with
every commit to the AOSP repository; we want the ability
to validate changes to the compiler for ChromeOS first.
This CL pins the GCC sources we pull from AOSP to the
most recent validated version of the sources.

BUG=chromium:608773
TEST='repo sync' works, fetches correct aosp GCC.
CQ-DEPEND=CL:*265825

Change-Id: I55e30214bab9fadfbc6c9cdae8783eb4cf3de838
Reviewed-on: https://chromium-review.googlesource.com/354460
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>