manifest: remove 'platform/system/hardware'
'platform/system/libhardware' was introduced by CL:266914 as a
dependency for 'platform/system/bt' for the development of Fluoride,
which has never been used on Chrome OS or is no longer in active
development. 'platform/system/libhardware' is only needed for
'platform/system/bt' and can thus be removed along with the latter.

BUG=chromium:673072
TEST=Remote trybot runs.

Change-Id: Ib740644a0f316eb6049d319f8afd33ec9a2079da
Reviewed-on: https://chromium-review.googlesource.com/422439
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>