Revert CL 438688 manifest: Update GCC pinned version.
The latest version of LLVM, when built with the updated version
of GCC, fails while building sys-libs/pam.   (LLVM updated
after this patch was tested.) We need to revert the GCC update
while this issue is debugged and fixed.

BUG=chromium:690639
CQ-DEPEND=CL:*326505
TEST=clang built with this version of GCC has the issue; clang
built with the old version of GCC does not.

Change-Id: I287396782330cd8f8a180f3ca9d3a9887df09bd6
Reviewed-on: https://chromium-review.googlesource.com/440468
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>