Add kernel-next
This is added to the kernel-next group, so it is not synced by
default. It point to the kernel-next repo instead of the kernel repo
so we don't clutter up the stable kernel repo with rebase artifacts. The
branching strategy is set to tot.

BUG=chromium:1011226, chromium:992138
TEST=repo sync src/third_party/kernel/next
git -C src/third_party/kernel/next show
Confirmed it points to 5.4 rebase

Change-Id: Iaf3904a3acf002cc096046e4ed935b2bfd978639
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/manifest/+/1893498
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>