manifest: Select the trembyle branch for blobs
We have decided not to commit the blobs to the master branch in the blobs
repo. The intention instead is to eventually build the binaries from
source.

In the meantime, point the blobs directory to the trembyle branch so that
ToT builders can build the firmware image.

BUG=chromium:1034094
TEST=repo sync -l; see that we get the new directories

Change-Id: I9effb87b03e8c62340b9f0a92e7000093c9d9ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/manifest/+/1980066
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>