manifest: Uprev puffin to the latest commit.
We need to get aosp/1240207 into Chrome OS but since dev-util/puffin
does not have a Chrome OS branch, the only way to do this cherrypick is
upreving the revision in the manifest itself and then upreving the
ebuild. This CL uprevs the public manifest.

BUG=chromium:1050869
TEST=sudo FEATURES=test emerge puffin

Cq-Depend: chrome-internal:2632641
Change-Id: If8978e64e6140567318e160a2c6a93083349e419
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/manifest/+/2072700
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>