Revert "whining: drop unused repo"
This reverts commit b0bd246e3114e4f80bfb2e1e1c878c3a984808c5.

Reason for revert:
The ebuild using this repo in ToT was deleted, but the manifest CL
landed before a new annealing snapshot was produced.  That has caused
the uprev process to fail as it still sees the old overlay/ebuild and
it wants this repo.  The annealing builders have been choking on an
unrelated infra flake since and haven't produced a new snapshot yet.
Lets re-add the repo to the manifest temporarily to see if it helps
unwedge the CQ until the annealing snapshot can get fixed.

Original change's description:
> whining: drop unused repo
>
> The ebuild referring to it has been removed.
>
> BUG=None
> TEST=CQ passes
>
> Change-Id: I01110d129d5b1f25d5dc5ba3801d228c8fd5d760
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/manifest/+/2176853
> Reviewed-by: Keith Haddow <haddowk@chromium.org>
> Commit-Queue: Mike Frysinger <vapier@chromium.org>
> Tested-by: Mike Frysinger <vapier@chromium.org>

BUG=chromium:1077539
TEST=CQ

Change-Id: I6e1b0352c1a917477854919f6237afe6cf75f6fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/manifest/+/2177107
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>